requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
exceptiongroup==1.1.3
idna==3.4
importlib-metadata==6.7.0
iniconfig==2.0.0
packaging==23.1
pluggy==1.2.0
PyMySQL==1.1.0
pytest==7.4.0
PyYAML==6.0.1
requests==2.31.0
tomli==2.0.1
typing-extensions==4.7.1
urllib3==2.0.4
zipp==3.15.0
